moved version module to separate file
